+++ django_admin_logs/settings.py
@@ -8,8 +8,8 @@ from django.conf import settings
 # If disabled, no log entries are created or displayed in the admin section.
 DJANGO_ADMIN_LOGS_ENABLED = getattr(settings, "DJANGO_ADMIN_LOGS_ENABLED", True)
 
-# Determines whether admin logs are deletable (defaults to False).
-# If enabled, non supers users will still require the delete_logentry permission.
+# Determines whether admin logs can be deleted (defaults to False).
+# If enabled, non-superusers will still require the delete_logentry permission.
 DJANGO_ADMIN_LOGS_DELETABLE = getattr(settings, "DJANGO_ADMIN_LOGS_DELETABLE", False)
 
 # Determines whether to ignore (not log) CHANGE actions where no changes were made